Added readme file.
Explained the initial motivation, what herri is, what its dependencies are.
Will still need to beef up the dependencies, and probably beef up the description
of the off screen stuff. Having said that, I've put a lin, to the awesome
vimeo video in there, because that explains it much better than I ever could.