Initial attempt at Vagrant support.
Doesn't utilise puppet or chef, due to my lack of experience with them.
Rather, it just executes a bunch of shell commands to provision the VM.

I also started moving the db_population folder into a subfolder of "server/setup".
The goal is to separate all of the VM provisioning stuff into there eventually.