Serve LGA geometry from django. Removed need for geoserver completely.
This was the final change required for the geoserver dependency to
be removed.

The view now also only requests data for the current part of the map
it is viewing (rather than all of Australia). It also specifies a
simplification level to return less simple geometry when zoomed out.

There were some troubles with performance in terms of:
 * Querying data from PostGIS (fast)
 * Using vectorformats to serialize to geoJSON, which requires:
   - Reading from PostGIS into internal django representations (slow)
   - Serializing that internal representation (fast enough)

In order to address this issue, this commit does a few things with cachecs:
 * Asks the browser to cache the result (with varying success
   between browsers - mostly because they seem to treat cache
   requests for JSON data loaded through AJAX differently from
   pages loaded over regular requests).
 * Use the DJango cache to cache results of the view which
   returns geometry (much more succesful, because the slow
   converting to JSON explained above is only done once).

For the caching to work, it is important to ensure that the requests
to the server for geometry (of a certain region, depending on the map
view) are not all completely different. Otherwise, caching would be
of no use. To achieve this, each time geometry is requested, it adds
a bit of padding to make the requested bounding box large. Then, it
rounds it off to an evenish number. That way, e.g. both 5.8 and 5.763
will result in the same request (e.g. "6.0").

Finally, the LGA shape files needed to be processed to remove features
with no geometry, as they make geodjango fall over. See prepare.sh
for details.