Proper quantiles + normalisation to population.
Additionally,

 * Saving models now respects RESTRICT_TO_STATE setting
 * Quantiles are saved against AttributeModel django model
 * JSON data used by JavaScript is no longer loaded over AJAX
 * Completely removed need for geoserver database

The JSON data was originally loaded over ajax because the client was
not being developed as part of the django app. In the context of the
hackathon, it made it easier to get stuff built. However, now that it
is being served as a django template, we can dump the data in the
template and thus avoid another round trip to the server. This is the
case with:
 * AttributeModel details
 * AttributeModel results (e.g. values for colourising the regions)
 * Attributes (in the new model view)