Changed colours to red, and exaggerate top three quantiles.
Fred mentioned that at present, we don't particularly care about
the people in the lower quantiles. By definition, they are less
interesting to us. Therefore we colour both the lowest quantiles
the same to make them fade into the background a bit.

In addition, the colours were changed to green, because often the
index will be modelling some group of either vulnerable people
or those requiring a service which don't have it. Green was a
very happy-go-lucky-colour for this job.

I kept the green colours in JavaScript to make it easy to swap back
though. Maybe it can be made into a django setting later (MAP_COLOUR)