Option to highlight the 5 highest scoring LGA's on the map.
The idea is that people look at the map for one reason - to find the
places with the highest scores on the index model. As they will not
neccesarily be able to distinguish between the LGA's in the highest,
quantile, this feature allows them to quickly jump their attention
to the relevant area.