Added labels to regions.
The labels are added for every region, and then culled back depending on
whether they overlap with other layers. Hence, the regions earlier in
the returned geoJSON will get priority. We can probably tweak this
algorithm later, as well as the symbology for labels, but for now it
gets the job done.