Setting for mapbox basemaps.
If you change local_settings.py and add a mapbox id in the setting
"MAPBOX_BASEMAP", then it will use that map for the basemap.
For example, "MAPBOX_BASEMAP = 'blah.2f03j'" - or whatever the format
for a mapbox account/map is. For testing I used mapbox id's from the
leaflet website, but don't want to commit them to our repository.