Added install path to wsgi.py. Fixed "create_or_update_autism_model()".
The wsgi.py now hard codes the install path to /opt/herri. This can
be changed in the future, but it is a requirement for now so that
apache can be setup correctly.

As for the load.create_or_update_autism_model(), this commit adds
optional verbose debugging to the AttributeModel.recalculate_index()
method. It also adds a special case for when a model only has one
attribute, and that attribute is total population. This stuffs up
the normalisation, so we don't normalise in this case.