More work on data importing, specifically data from ABS "Table Builder"
The table builder file is one that Fred sent to me. Not quite sure how
or where it comes from, but I'm not committing it here. This is for the
same reason as not commiting other census data (don't have the rights).

It is a bit weird how the import runs now - it firstly nukes the
lga_attributes_all table completely. Then, it creates it with only
the region_id and tot_p_p fields. After this, we iterate over each of
the columns in the table builder CSV and add each field to the table.
Then the data is imported for each row. Finally, we hijack the "b01"
table from the census data pack import, in order to find out the
tot_p_p value (which wasn't in the table builder csv file).