Updated attributes import sql. Better docs for configuring census data.
The admin form which allows adding of attributes now has a drop down
select list for available census attributes. This saves us having to
manually validate that each one which gest added is the real deal.

However, this is still not safe. The real solution is to not have data
stored as great big rows with hundreds of attributes. Rather, it would
be better and safer to store each value as a tuple with: attribute +
value. This is a task for the future though.